Set BeamOptikDLL truth value from initialization state
Since bc208c7 BeamOptikDLL objects do NOT represent interface instances
anymore but rather the underlying loaded shared object which may or may
be initialized with a call to GetInterfaceInstance. It is natural to use
the truth value in boolean contexts to represent this initialization
state. The current check in the hit.online_control.madgui plugin is also
based on this truth value and needs this patch to keep working.