Revert "Auto-set VAcc from sequence"
This reverts commit f65a115d83edfb3fc7452d426e7ce9b10bedcdd8.

As long as this is not more sophisticated, it is rather useless anyway
and can be pretty annoying that the VAcc/MEFI is reset by the program
every time the sequence changes (e.g. if you changed the MEFI
combination before changing the sequence).

We might come back to this once we have a more reliably method that will
at least not destroy the user settings.